Moved and renamed CheckInOutRequest
